Reorgnize and deduplicate files
