--- CONTRIBUTORS.md
@@ -1,70 +1,0 @@
-<!--- Licensed to the Apache Software Foundation (ASF) under one -->
-<!--- or more contributor license agreements.  See the NOTICE file -->
-<!--- distributed with this work for additional information -->
-<!--- regarding copyright ownership.  The ASF licenses this file -->
-<!--- to you under the Apache License, Version 2.0 (the -->
-<!--- "License"); you may not use this file except in compliance -->
-<!--- with the License.  You may obtain a copy of the License at -->
-
-<!---   http://www.apache.org/licenses/LICENSE-2.0 -->
-
-<!--- Unless required by applicable law or agreed to in writing, -->
-<!--- software distributed under the License is distributed on an -->
-<!--- "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY -->
-<!--- KIND, either express or implied.  See the License for the -->
-<!--- specific language governing permissions and limitations -->
-<!--- under the License. -->
-
-# How to Contribute
-
-We'd love to accept your patches and contributions to this project. There are
-just a few small guidelines you need to follow.
-
-## Contributor License Agreement
-
-Contributions to this project must be accompanied by a Contributor License
-Agreement. You (or your employer) retain the copyright to your contribution;
-this simply gives us permission to use and redistribute your contributions as
-part of the project.
-
-You generally only need to submit a CLA once, so if you've already submitted one
-(even if it was for a different project), you probably don't need to do it
-again.
-
-## Changes Accepted
-
-Please file issues before doing substantial work; this will ensure that others
-don't duplicate the work and that there's a chance to discuss any design issues.
-
-Changes only tweaking style are unlikely to be accepted unless they are applied
-consistently across the project. Most of the code style is derived from the
-[Google Style Guides](http://google.github.io/styleguide/) for the appropriate
-language and is generally not something we accept changes on (as clang-format
-and clang-tidy handle that for us). The compiler portion of the project follows
-[MLIR style](https://mlir.llvm.org/getting_started/DeveloperGuide/#style-guide).
-Improvements to code structure and clarity are welcome but please file issues to
-track such work first.
-
-## AUTHORS file
-
-If you would like to receive additional recognition for your contribution, you
-may add yourself (or your organization) to the AUTHORS file. This keeps track of
-those who have made significant contributions to the project. Please add the
-entity who owns the copyright for your contribution. The source control history
-remains the most accurate source for individual contributions.
-
-##  Pull Requests
-We actively welcome your pull requests.
-
-1. Fork the repo and create your branch from `main`.
-2. If you've added code that should be tested, add tests.
-3. If you've changed APIs, update the documentation.
-4. Ensure the test suite passes.
-5. Make sure your code lints.
-6. If you haven't already, complete the Contributor License Agreement ("CLA").
-
-## Issues
-
-We use GitHub issues to track public bugs. Please ensure your description is
-clear and has sufficient instructions to be able to reproduce the issue.
-
