+++ python/flux/__init__.py
@@ -15,6 +15,9 @@
 #
 ################################################################################
 
+# version info
+__version__ = "1.0.0"
+
 from .cpp_mod import *
 from .ag_gemm import *
